move VirtualDirectoryManager  to a new ts file
